check and update documentation too
checks and fixes the integrity of the change log section and hash table,
and fixes a variety of foreseeable problems.
In pack zip section, inserts the MD5 into the non-zipped documentation;
later checks will add a line for the next version.
Potential problem if version changes after running and I don't manually
fix, but it'll still make syntactic sense and I can't be bothered fixing
that now.